fix: generate after template
